requirements.txt
streamlit>=1.32
google-cloud-bigquery>=3.17
db-dtypes>=0.5
pyarrow>=14.0
pandas>=2.0
jinja2>=3.1
playwright>=1.44.0
matplotlib>=3.8
pyyaml>=6.0
markdown-it-py>=3.0
google-auth>=2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
pytest>=8.0
